dockercompose updated, we can now work with layers. Yay.
