+++ calendify/src/presentationLayer/login-router.js
@@ -1,23 +1,18 @@
-// const express = require('express')
-// const expressSession = require('express-session')
-// const bcrypt = require('bcryptjs')
-// const router = express.Router()
+const bcrypt = require('bcryptjs')
 
-
-
-// const enteredUsername = request.body.username
-// const enteredPassword = request.body.password
-
-// router.get("/", function(req, res){
+function login(){
+    //TODO: Re-structure login to appropiate layer, also only admin atm. Needs to be fixed as well.
+    const enteredUsername = request.body.username
+    const enteredPassword = request.body.password
   
-// })
-// // const checkPassword = bcrypt.compareSync(enteredPassword, adminPassword)
-
-// // if (enteredUsername == adminUsername && checkPassword) {
-// //   //Login
-// //   request.session.isLoggedIn = true
-// //   response.redirect("/")
-// // } else {
-// //   //display error message to user
-// // }
-
+    const checkPassword = bcrypt.compareSync(enteredPassword, adminPassword)
+  
+    if (enteredUsername == adminUsername && checkPassword) {
+      //Login
+      request.session.isLoggedIn = true
+      console.log("never gonna give you up...")
+      response.redirect("/")
+    } else {
+      //TODO:display error message to user
+    }
+}
