+++ docker-compose.yml
@@ -5,4 +5,4 @@ services:
     ports: 
       - 8080:8080
     volumes: 
-      - ./calendify/src:/calendifyDock/src/
+      - ./calendify/src:/calendifyDock/src/presentationLayer/
